+++ src/evaluator/commands/analyse/analyse.py
@@ -76,9 +76,9 @@ def run_pipeline(
         lg.warning(f"Nothing saved to {out_file}.")
         return
     lg.debug(f"analyse | Saving output CSV ({out_file.name})...")
-    analyse_df = saveResultsCSV(analyse_results, out_file)
+    analyse_df = io.saveResultsCSV(analyse_results, out_file)
     lg.debug(f"analyse | Printing summary message...")
-    printSummaryMessage(results=analyse_df, nfiles=len(seg_files), startt=START_TIME, endt=END_TIME, out_path=out_file)
+    io.printSummaryMessage(results=analyse_df, nfiles=len(seg_files), startt=START_TIME, endt=END_TIME, out_path=out_file)
 
 # =========================
 # DEFINE FUNCTION: processSegmentation
@@ -191,29 +191,3 @@ def morphologicalClosure(binary_vol: numpy.ndarray):
     '''
     struc = ndimage.generate_binary_structure(3, 1)
     return ndimage.binary_closing(binary_vol, structure=struc, border_value=False)
-
-# =========================
-# DEFINE FUNCTION: saveResultsCSV
-# =========================
-def saveResultsCSV(analyse_results, out_path: Path):
-    '''
-    Convert list of result dictionaries to a pandas DataFrame and save to CSV.
-    '''
-    analyse_df = pandas.DataFrame(analyse_results)
-    analyse_df.to_csv(out_path, index=False)
-    lg.info(f"Analyse results saved to: {out_path}")
-    return analyse_df
-
-# =========================
-# DEFINE FUNCTION: printSummaryMessage
-# =========================
-def printSummaryMessage(results, nfiles: int, startt: datetime.datetime, endt: datetime.datetime, out_path: Path):
-    RUNTIME = endt - startt
-    print(f"\n[bold]Pipeline run summary[/bold]")
-    print(f"- Runtime: {RUNTIME}")
-    print(f"- Segmentation files processed: {nfiles}")
-    print(f"- Segmentation files with EVs: {results['tomogram'].nunique()} ({(100 * results['tomogram'].nunique()) / nfiles:.1f}%)")
-    print(f"- EVs processed: {len(results)}")
-    print(f"- Number of enclosed EVs: {results['is_enclosed'].sum()} ({100 * results['is_enclosed'].mean():.1f}%)")
-    print(f"- Equivalent diameters: {results['equiv_diameter_nm'].mean():.1f} ± {results['equiv_diameter_nm'].std():.1f} nm (mean ± SD)")
-    print(f"Results saved to: {out_path}\n")
